requirements.txt
openai-whisper
customtkinter
ffmpeg-python
